pyproject.toml
[project]
name = "jlcpcb-tool"
version = "0.1.0"
description = "CLI tool for researching JLCPCB/LCSC electronic components"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "requests>=2.28",
    "pyyaml>=6.0",
    "tabulate>=0.9",
]

[project.scripts]
jlcpcb = "jlcpcb_tool.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jlcpcb_tool"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-recording>=0.13",
]
